fix(mac-vibes): clean up settings card structure
- split Vibes AI install controls into their own settings card file
- move Vibe examples into a dedicated style card section
- scope settings copy to the views that own it
- align the Vibes style icon container with the shared settings row treatment
