+++ macOS/Views/Settings/SettingsVibesCard.swift
@@ -7,11 +7,6 @@ struct SettingsVibesCard: View {
     let downloadAction: () -> Void
     let repairAction: () -> Void
 
-    @State private var isVibeExamplesExpanded = false
-    @State private var vibeExamplesExpandedContentHeight: CGFloat = 0
-
-    private static let sectionExpansionAnimation = Animation.spring(response: 0.42, dampingFraction: 0.84)
-
     var body: some View {
         VStack(alignment: .leading, spacing: 16) {
             SettingsCard {
@@ -25,41 +20,11 @@ struct SettingsVibesCard: View {
                     Divider()
                         .overlay(.white.opacity(0.22))
 
-                    HStack(alignment: .top, spacing: 12) {
-                        Text(matrix.displayedSelectedVibe.description)
-                            .font(.appFont(15, variant: .light))
-                            .foregroundStyle(.white.opacity(0.7))
-                            .frame(maxWidth: .infinity, alignment: .leading)
-
-                        Button {
-                            withAnimation(Self.sectionExpansionAnimation) {
-                                isVibeExamplesExpanded.toggle()
-                            }
-                        } label: {
-                            Image(systemName: isVibeExamplesExpanded ? "chevron.down" : "chevron.right")
-                                .font(.system(size: 28, weight: .heavy))
-                                .foregroundStyle(.yellow)
-                                .frame(width: 56, height: 56)
-                                .contentShape(Rectangle())
-                        }
-                        .buttonStyle(.plain)
-                        .accessibilityLabel(
-                            isVibeExamplesExpanded
-                            ? MacVibesSettingsCopy.hideExamplesAccessibilityLabel
-                            : MacVibesSettingsCopy.showExamplesAccessibilityLabel
-                        )
-                    }
-
-                    vibeExamplesExpandedContent
-                        .frame(height: isVibeExamplesExpanded ? vibeExamplesExpandedContentHeight : 0, alignment: .top)
-                        .clipped()
-                        .allowsHitTesting(isVibeExamplesExpanded)
-                        .accessibilityHidden(!isVibeExamplesExpanded)
-                        .background(alignment: .top) {
-                            if isVibeExamplesExpanded || vibeExamplesExpandedContentHeight == 0 {
-                                vibeExamplesExpandedContentMeasurement
-                            }
-                        }
+                    SettingsVibesExamplesSection(
+                        selectedVibe: $selectedVibe,
+                        displayedSelectedVibe: matrix.displayedSelectedVibe,
+                        isSelectionEnabled: matrix.showsVibeSelector
+                    )
                 }
             }
 
@@ -67,7 +32,7 @@ struct SettingsVibesCard: View {
                 Spacer()
                 TipItem(
                     icon: "keyboard",
-                    text: MacVibesSettingsCopy.triggerTip
+                    text: SettingsVibesCardCopy.triggerTip
                 )
                 Spacer()
             }
@@ -77,20 +42,20 @@ struct SettingsVibesCard: View {
     private var headerContent: some View {
         HStack(alignment: .top, spacing: 12) {
             ZStack {
-                Circle()
-                    .fill(MacAppTheme.accent.opacity(0.4))
-                    .frame(width: 32, height: 32)
+                RoundedRectangle(cornerRadius: 12)
+                    .fill(MacAppTheme.iconFill)
+                    .frame(width: 44, height: 44)
 
                 Image("vibes-logo")
                     .resizable()
                     .renderingMode(.template)
                     .scaledToFit()
                     .foregroundStyle(MacAppTheme.accent)
-                    .frame(width: 18, height: 18)
+                    .frame(width: 24, height: 24)
             }
 
             VStack(alignment: .leading, spacing: 4) {
-                Text(MacVibesSettingsCopy.cardTitle)
+                Text(SettingsVibesCardCopy.cardTitle)
                     .font(.appFont(18))
                     .foregroundStyle(.white)
             }
@@ -113,14 +78,14 @@ struct SettingsVibesCard: View {
         switch matrix.cardControl {
         case .download:
             AppActionButton(
-                title: MacVibesSettingsCopy.downloadAction,
+                title: SettingsVibesCardCopy.downloadAction,
                 style: .primary,
                 minWidth: 84,
                 action: downloadAction
             )
         case .repair:
             AppActionButton(
-                title: MacVibesSettingsCopy.repairAction,
+                title: SettingsVibesCardCopy.repairAction,
                 style: .primary,
                 minWidth: 84,
                 action: repairAction
@@ -136,7 +101,7 @@ struct SettingsVibesCard: View {
             }
             .pickerStyle(.menu)
             .labelsHidden()
-            .accessibilityLabel(MacVibesSettingsCopy.pickerAccessibilityLabel)
+            .accessibilityLabel(SettingsVibesCardCopy.pickerAccessibilityLabel)
         }
     }
 
@@ -164,9 +129,9 @@ struct SettingsVibesCard: View {
     private var progressBadgeTitle: String {
         switch matrix.mainCardContent {
         case .installing:
-            return MacVibesSettingsCopy.installingBadge
+            return SettingsVibesCardCopy.installingBadge
         case .downloading:
-            return MacVibesSettingsCopy.downloadingBadge
+            return SettingsVibesCardCopy.downloadingBadge
         case .downloadRequired, .installFailed, .selectedVibe:
             return ""
         }
@@ -175,15 +140,15 @@ struct SettingsVibesCard: View {
     private var statusText: String {
         switch matrix.mainCardContent {
         case .downloadRequired:
-            return MacVibesSettingsCopy.downloadRequiredStatus
+            return SettingsVibesCardCopy.downloadRequiredStatus
         case .downloading:
-            return MacVibesSettingsCopy.downloadingStatus
+            return SettingsVibesCardCopy.downloadingStatus
         case .installing:
-            return MacVibesSettingsCopy.installingStatus
+            return SettingsVibesCardCopy.installingStatus
         case .installFailed:
-            return MacVibesSettingsCopy.installFailedStatus
+            return SettingsVibesCardCopy.installFailedStatus
         case .selectedVibe:
-            return MacVibesSettingsCopy.readyStatus
+            return SettingsVibesCardCopy.readyStatus
         }
     }
 
@@ -199,102 +164,10 @@ struct SettingsVibesCard: View {
             return true
         }
     }
-
-    private var vibeExamplesExpandedContent: some View {
-        VStack(alignment: .leading, spacing: 16) {
-            Divider()
-                .overlay(.white.opacity(0.22))
-
-            VStack(alignment: .leading, spacing: 14) {
-                ForEach(Array(vibeExamples.enumerated()), id: \.element.style) { index, example in
-                    vibeExampleRow(example)
-
-                    if index < vibeExamples.count - 1 {
-                        Divider()
-                            .overlay(.white.opacity(0.22))
-                            .padding(.leading, 12)
-                            .padding(.trailing, 12)
-                    }
-                }
-            }
-        }
-    }
-
-    private var vibeExamplesExpandedContentMeasurement: some View {
-        vibeExamplesExpandedContent
-            .fixedSize(horizontal: false, vertical: true)
-            .hidden()
-            .allowsHitTesting(false)
-            .accessibilityHidden(true)
-            .background(
-                GeometryReader { geometry in
-                    Color.clear
-                        .onAppear {
-                            updateVibeExamplesExpandedContentHeight(geometry.size.height)
-                        }
-                        .onChange(of: geometry.size.height) { newHeight in
-                            updateVibeExamplesExpandedContentHeight(newHeight)
-                        }
-                }
-            )
-    }
-
-    private func vibeExampleRow(_ example: VibeExample) -> some View {
-        Button {
-            guard matrix.showsVibeSelector else { return }
-            selectedVibe = example.style
-        } label: {
-            HStack(alignment: .center, spacing: 8) {
-                VStack(alignment: .leading, spacing: 6) {
-                    Text(example.style.displayName)
-                        .font(.appFont(17))
-                        .foregroundStyle(.yellow)
-                        .lineLimit(1)
-                        .fixedSize(horizontal: true, vertical: false)
-
-                    Text(example.text)
-                        .font(.appFont(15, variant: .light))
-                        .foregroundStyle(.white)
-                        .frame(maxWidth: .infinity, alignment: .leading)
-                        .fixedSize(horizontal: false, vertical: true)
-                }
-                .frame(maxWidth: .infinity, alignment: .leading)
-
-                Image(systemName: example.style == matrix.displayedSelectedVibe ? "checkmark.circle.fill" : "checkmark.circle.dotted")
-                    .font(.system(size: 15, weight: .medium))
-                    .foregroundStyle(example.style == matrix.displayedSelectedVibe ? .green : .white)
-            }
-            .padding(.leading, 10)
-            .padding(.trailing, 15)
-            .frame(maxWidth: .infinity, alignment: .leading)
-            .contentShape(Rectangle())
-        }
-        .frame(maxWidth: .infinity, alignment: .leading)
-        .buttonStyle(.plain)
-    }
-
-    private func updateVibeExamplesExpandedContentHeight(_ newHeight: CGFloat) {
-        guard abs(vibeExamplesExpandedContentHeight - newHeight) > 0.5 else { return }
-        vibeExamplesExpandedContentHeight = newHeight
-    }
-
-    private var vibeExamples: [VibeExample] {
-        StyleRewriteStyle.allCases.map { style in
-            VibeExample(style: style, text: style.exampleText)
-        }
-    }
-
-    private struct VibeExample: Hashable {
-        let style: StyleRewriteStyle
-        let text: String
-    }
 }
 
-enum MacVibesSettingsCopy {
+enum SettingsVibesCardCopy {
     static let cardTitle = "KeyVox Vibes"
-    static let aiCardTitle = "KeyVox Vibes AI"
-    static let aiReadyStatus = "KeyVox Vibes AI is installed and ready."
-    static let aiDownloadRequiredStatus = "Install Vibes AI first (~491 MB), then you can use KeyVox Vibes."
     static let pickerAccessibilityLabel = "KeyVox Vibes"
     static let downloadRequiredStatus = "Install Vibes AI first (~491 MB), then you can use KeyVox Vibes."
     static let downloadingStatus = "Downloading KeyVox Vibes AI."
@@ -305,178 +178,5 @@ enum MacVibesSettingsCopy {
     static let installingBadge = "Installing"
     static let downloadAction = "Download"
     static let repairAction = "Repair"
-    static let deleteAction = "Delete"
-    static let showExamplesAccessibilityLabel = "Show vibe examples"
-    static let hideExamplesAccessibilityLabel = "Hide vibe examples"
     static let triggerTip = "Tap the trigger key to apply / undo the current Vibe. Double-tap to cycle Vibes."
-    static let triggerKeyInteractionsTitle = "Trigger Key Interactions"
-    static let triggerKeyInteractionsSubtitle = "Tap to apply / undo Vibes. Double-tap to cycle Vibes."
-    static let triggerKeyInteractionsAccessibilityLabel = "Vibes trigger key interactions"
-    static let deleteConfirmationTitle = "Delete KeyVox Vibes AI?"
-    static let deleteConfirmationMessage = "KeyVox Vibes AI will be removed from this Mac."
-}
-
-struct SettingsVibesAIInstallCard: View {
-    @Binding var triggerKeyInteractionsEnabled: Bool
-    let installState: MacLocalRewriteModelInstallState
-    let downloadAction: () -> Void
-    let repairAction: () -> Void
-    let deleteAction: () -> Void
-
-    var body: some View {
-        SettingsCard {
-            VStack(spacing: 16) {
-                SettingsRow(
-                    assetIcon: "vibes-logo",
-                    title: MacVibesSettingsCopy.aiCardTitle,
-                    subtitle: subtitleText
-                ) {
-                    actionControl
-                }
-
-                Divider()
-                    .background(Color.white.opacity(0.22))
-
-                triggerKeyInteractionsToggle
-
-                if shouldShowInstallContent {
-                    Divider()
-                        .background(Color.white.opacity(0.22))
-
-                    installContent
-                }
-            }
-        }
-    }
-
-    private var triggerKeyInteractionsToggle: some View {
-        HStack(alignment: .center, spacing: 12) {
-            VStack(alignment: .leading, spacing: 4) {
-                Text(MacVibesSettingsCopy.triggerKeyInteractionsTitle)
-                    .font(.appFont(14))
-                    .foregroundStyle(.white)
-
-                Text(MacVibesSettingsCopy.triggerKeyInteractionsSubtitle)
-                    .font(.appFont(12, variant: .light))
-                    .foregroundColor(.secondary)
-                    .fixedSize(horizontal: false, vertical: true)
-            }
-            .frame(maxWidth: .infinity, alignment: .leading)
-
-            Toggle("", isOn: $triggerKeyInteractionsEnabled)
-                .toggleStyle(.switch)
-                .labelsHidden()
-                .accessibilityLabel(MacVibesSettingsCopy.triggerKeyInteractionsAccessibilityLabel)
-        }
-    }
-
-    @ViewBuilder
-    private var actionControl: some View {
-        switch installState {
-        case .notInstalled:
-            AppActionButton(
-                title: MacVibesSettingsCopy.downloadAction,
-                style: .primary,
-                minWidth: 84,
-                action: downloadAction
-            )
-        case .ready:
-            AppActionButton(
-                title: MacVibesSettingsCopy.deleteAction,
-                style: .destructive,
-                minWidth: 84,
-                action: deleteAction
-            )
-        case .failed:
-            AppActionButton(
-                title: MacVibesSettingsCopy.repairAction,
-                style: .primary,
-                minWidth: 84,
-                action: repairAction
-            )
-        case .downloading, .installing:
-            EmptyView()
-        }
-    }
-
-    private var installContent: some View {
-        VStack(alignment: .leading, spacing: 8) {
-            HStack(alignment: .center, spacing: 12) {
-                Text(statusText)
-                    .font(.appFont(12, variant: .light))
-                    .foregroundColor(.secondary)
-                    .frame(maxWidth: .infinity, alignment: .leading)
-
-                if let percentageText {
-                    Text(percentageText)
-                        .font(.appFont(12))
-                        .foregroundStyle(.yellow)
-                }
-            }
-
-            if let errorMessage {
-                Text(errorMessage)
-                    .font(.appFont(10))
-                    .foregroundColor(.red)
-                    .frame(maxWidth: .infinity, alignment: .leading)
-            }
-
-            if let progress {
-                ModelDownloadProgress(progress: progress)
-            }
-        }
-    }
-
-    private var subtitleText: String {
-        switch installState {
-        case .ready:
-            return MacVibesSettingsCopy.aiReadyStatus
-        case .notInstalled, .downloading, .installing, .failed:
-            return MacVibesSettingsCopy.aiDownloadRequiredStatus
-        }
-    }
-
-    private var statusText: String {
-        switch installState {
-        case .downloading:
-            return MacVibesSettingsCopy.downloadingStatus
-        case .installing:
-            return MacVibesSettingsCopy.installingStatus
-        case .failed:
-            return MacVibesSettingsCopy.installFailedStatus
-        case .notInstalled, .ready:
-            return ""
-        }
-    }
-
-    private var progress: Double? {
-        switch installState {
-        case .downloading(let progress), .installing(let progress):
-            return progress
-        case .notInstalled, .ready, .failed:
-            return nil
-        }
-    }
-
-    private var percentageText: String? {
-        guard let progress else { return nil }
-        return "\(Int((progress * 100).rounded()))%"
-    }
-
-    private var errorMessage: String? {
-        if case .failed(let message) = installState {
-            return message
-        }
-
-        return nil
-    }
-
-    private var shouldShowInstallContent: Bool {
-        switch installState {
-        case .downloading, .installing, .failed:
-            return true
-        case .notInstalled, .ready:
-            return false
-        }
-    }
 }
